taxonomy: qualify iOS devices with DHCP OS.
The percentage of iOS devices in the results is supposed to
be high owing to their enormous popularity... but even with
that we have a suspiciously high percentage of iPhones and
iPads. It is possible that some other devices are aliasing
the Wifi signatures of the iOS devices. Qualify these
signatures with the DHCP signature for iOS.

Also:
+ add OUIs for Google (Chromecast), Samsung, and HTC.
+ add alternate signatures for Chromebook Pixel 2, Nexus 7,
    iPad Air 2.
+ add signature for iPod Touch 5th generation
+ update to v2 signatures: Samsung Galaxy Note 3,
    Samsung Galaxy Note 4.
+ Correct 5GHz signature for Nexus 9, and remove the OUI qualifier.
    The Wifi signature contains a WPS model name, which is
    sufficiently specific.
+ make wtax immediately print known devices when it starts.

Change-Id: I7301018419757bf0af0b1681ac7b602619513b20