taxonomy: construct a test from pcaps.
Add a collection of pcap files collected from various devices.
All of these devices are either test devices from Google, or
belong to me personally.

Add a utility to extract a signature from a pcap file, using
the same code which hostapd uses for signatures.

Construct a test that all pcap files resolve correctly, to catch
if we introduce accidental aliasing or breakage in a database update.

The v2 signatures for these devices passed, but take the opportunity
to update them to v3 while here.

Change-Id: I15b01fa97c024c8d0270371e1015a65c695601bf