taxonomy: v3 for Kindle Fire, iPod 5G
Update signatures for Amazon Kindle Fire (2011 edition)
and iPod Touch 5th gen to v3 format.

Add a signature for one variant of Samsung Galaxy Tab (1st gen)
with txpow values which are distinctive.

Add an alternate Samsung Galaxy S4 signature.

Add pcaps for a few more phones which are not yet known to be
common enough to add to the database.

Add OUIs for Amazon, HTC, Samsung.

Change-Id: I80437ac00fe832771d79321164e7687765dc9bfc