taxonomy: add anonymize_pcap
Rewrites a pcap file to obscure MAC addresses and SSIDs.

We gather pcaps for the taxonomy unit test, checking them into
testdata/pcaps. Some of the pcaps are from people's
personal devices or from their Wifi networks at home. As we
publish this code and these pcaps, this utility obscures the
MAC address and SSID of people's personal devices before
checking them in.

Change-Id: I743af0231c3e81aec4aa360e69b9e83fecd2475c