Remove DHCP qualification from iOS 10 devices.
As noted in the paper (https://arxiv.org/abs/1608.01725) we
want to ensure signatures as sufficiently unique by qualifying
them with either a specific WPS name, an OUI, a DHCP options
signature, or DHCP name.

We'll add a fifth suitable qualification: presence of
the Apple vendor-specific IE added in iOS10. This is
sufficient to ensure it uniquely matches Apple devices
without needing to further check the DHCP options.

  Tag: Vendor Specific: Apple
    Tag Number: Vendor Specific (221)
    Tag length: 11
    OUI: 00-17-f2 (Apple)
    Vendor Specific OUI Type: 10
    Vendor Specific Data: 0a00010400000000

If Apple ever opens or licenses whatever this is for other
manufacturers to use, which we consider to be highly unlikely,
we'll reconsider this decision.