Add support for MA-M blocks.
OUIs are 24 bit identifiers assigne to a manufacturer.
Concatenated with 24 bits supplied by the manufacturer to
make a 48 bit MAC address, they therefore allow 2^24 = 16M
station addresses per OUI.

MA-M blocks are 28 bit identifiers, which allow 2^20 = 1M
stations to be manufactured. They are somewhat cheaper than
a whole OUI.

There is no portion of the MAC address space carved off for use
as MA-M blocks. The IEEE appears to consume a 24 bit OUI and carve
it up into 16 MA-M blocks (which is wise, it doesn't presuppose
how popular MA-M blocks will be).

Therefore, we first search the leading 28 bits of a MAC address
to see if it is an MA-M, then we check the 24 bits to see if it
is an OUI.