Experiment with gperf & remove duplicates.
Just an experiment with using gperf to generate a perfect
hash of the signatures. Unfortunately gperf itself isn't
going to work for this: gperf generates a hash function using
only the first 255 characters of the key, and 20 of the
signatures are not unique in their first 255 characters.
This is just a checkpoint, we'll see where it goes from here.

Also remove three signatures which were total duplicates,
identified while working with gperf.