Fix pcap snaplen to be less than MAXIMUM_SNAPLEN.
In 10.2015,
https://github.com/the-tcpdump-group/libpcap/commit/f983e075fbef40fe12323c4dd8f85c88eaf0f789
added a check of the pcap file's snaplen field, that it not exceed the
MAXIMUM_SNAPLEN of 256k or it would print an error and exit.

In 5/2017,
https://github.com/the-tcpdump-group/libpcap/commit/2be9c29d45fb1fab8e9549342a30c160b7dea3e1
softened the check somewhat, making it print and error and truncate the
length it would allow to MAXIMUM_SNAPLEN but not exit.

As of 2/2018, there are a number of Linux distributions which contain the
first but not the second commit, so libpcap will exit if it finds a snaplen
bigger than 256k.

Many pcaps in this repo were captured using Apple's Wi-Fi Diagnostics, which
sets the snaplen to 512k. There are no packets in any pcap which are more
than a few hundred bytes, it is just the snaplen in the header.

Add a new program to limit snaplen to 256k, in fix_pcap_snaplen, and use
it to fix all of the files in the repository.